% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-registry.R
\name{list_sources}
\alias{list_sources}
\alias{list_learners}
\title{List registered sources and learners}
\usage{
list_sources()

list_learners()
}
\value{
A data frame describing what is currently registered.
}
\description{
List registered sources and learners
}
\examples{
list_sources()
list_learners()
}
\seealso{
\code{\link[=register_source]{register_source()}}, \code{\link[=register_learner]{register_learner()}}
}
